+++ Mindkeepr/models/__init__.py
@@ -1,3 +1,5 @@
 from .category import *
 from .location import *
 from .models import *
+from .stock_repartition import StockRepartition
+from .project import Project
