+++ Mindkeepr/models/elements/__init__.py
@@ -1,2 +1,5 @@
 from .element import Element
 from .machine import Machine
+from .component import Component
+from .tool import Tool
+from .book import Book
