+++ Mindkeepr/models/elements/book.py
@@ -1,0 +1,8 @@
+
+from . import element
+
+class Book(element.Element):
+    """ Book """
+    # TODO Add ISBN/Barcode
+    def is_unique(self):
+        return True
