+++ Mindkeepr/models/elements/component.py
@@ -1,0 +1,10 @@
+
+from django.db import models
+
+from . import mixins
+from . import Element
+
+class Component(mixins.Consumable,Element):
+    """ Electronic component """
+    datasheet = models.FileField(upload_to='datasheet', blank=True, null=True)
+    pass
