+++ Mindkeepr/models/elements/tool.py
@@ -1,0 +1,6 @@
+
+from . import element
+
+class Tool(element.Element):
+    """ Tool """
+    pass
