+++ Mindkeepr/models/events/borrow_event.py
@@ -3,7 +3,7 @@ from django.db import models
 from datetime import date
 
 from . import Event
-
+from . import ReturnEvent
 
 class BorrowEvent(Event):
     quantity = models.IntegerField("Quantity", null=False, blank=False)
