+++ Mindkeepr/models/events/return_event.py
@@ -1,5 +1,4 @@
 from django.db import models
-from datetime import date
 
 from . import Event
