+++ Mindkeepr/views/__init__.py
@@ -1,0 +1,3 @@
+from .views import *
+from .mixins import *
+from .events import *
