+++ Mindkeepr/views/elements/__init__.py
@@ -1,0 +1,5 @@
+from .element_view import *
+from .book_view import *
+from .component_view import *
+from .machine_view import *
+from .tool_view import *
