+++ Mindkeepr/views/events/__init__.py
@@ -1,0 +1,9 @@
+from .event_view import *
+from .borrowing_view import *
+from .buy_view import *
+from .sell_view import *
+from .incident_view import *
+from .reserve_view import *
+from .maintenance_view import *
+from .consume_view import *
+from .move_view import *
