+++ Mindkeepr/views/events/move_view.py
@@ -1,0 +1,9 @@
+
+from . import EventViewModal
+from Mindkeepr.forms import MoveEventForm
+
+class MoveEventViewModal(EventViewModal):
+    template_name = 'events/move-event-detail-modal.html'
+    permission_required = "Mindkeepr.add_moveevent"
+    form_class = MoveEventForm
+    success_url = '/formmoveeventmodal'
