+++ MindkeeprMain/urls.py
@@ -80,15 +80,15 @@ urlpatterns = [
     path("element/print/remove/<int:pk>/<int:qty>",views.remove_from_print_list),
     path("element/print/all",views.print_print_list),
     path("printlist",views.print_list_disp),
-    path('formborroweventmodal', views.BorrowEventViewModal.as_view()),
-    path('formreturneventmodal', views.ReturnEventViewModal.as_view()),
-    path('formbuyeventmodal', views.BuyEventViewModal.as_view()),
-    path('formselleventmodal', views.SellEventViewModal.as_view()),
-    path('formuseeventmodal', views.UseEventViewModal.as_view()),
-    path('formmoveeventmodal', views.MoveEventViewModal.as_view()),
-    path('formunuseeventmodal', views.UnUseEventViewModal.as_view()),
-    path('formmaintenanceeventmodal', views.MaintenanceEventViewModal.as_view()),
-    path('formincidenteventmodal', views.IncidentEventViewModal.as_view()),
+    path('formborroweventmodal', views.events.BorrowEventViewModal.as_view()),
+    path('formreturneventmodal', views.events.ReturnEventViewModal.as_view()),
+    path('formbuyeventmodal', views.events.BuyEventViewModal.as_view()),
+    path('formselleventmodal', views.events.SellEventViewModal.as_view()),
+    path('formuseeventmodal', views.events.UseEventViewModal.as_view()),
+    path('formmoveeventmodal', views.events.MoveEventViewModal.as_view()),
+    path('formunuseeventmodal', views.events.UnUseEventViewModal.as_view()),
+    path('formmaintenanceeventmodal', views.events.MaintenanceEventViewModal.as_view()),
+    path('formincidenteventmodal', views.events.IncidentEventViewModal.as_view()),
     path('formmaintenanceeventmodal/<int:pk>', views.MaintenanceEventUpdateViewModal.as_view()),
     path('formconsumeeventmodal', views.ConsumeEventViewModal.as_view()),
     path('elements', views.elements,name="elements-list"),
